LayoutTests/fast/dom/Window/atob-btoa-expected.txt

PASS window.atob("YQ==") is "a"
PASS window.atob("YWI=") is "ab"
PASS window.atob("YWJj") is "abc"
PASS window.atob("YWJjZA==") is "abcd"
PASS window.atob("YWJjZGU=") is "abcde"
PASS window.atob("YWJjZGVm") is "abcdef"
PASS window.btoa("a") is "YQ=="
PASS window.btoa("ab") is "YWI="
PASS window.btoa("abc") is "YWJj"
PASS window.btoa("abcd") is "YWJjZA=="
PASS window.btoa("abcde") is "YWJjZGU="
PASS window.btoa("abcdef") is "YWJjZGVm"
PASS typeof window.btoa is "function"
PASS window.btoa() is "dW5kZWZpbmVk"
PASS window.btoa("") is ""
PASS window.btoa(null) is ""
PASS window.btoa(undefined) is "dW5kZWZpbmVk"
PASS window.btoa(window) is "W29iamVjdCBET01XaW5kb3dd"
PASS window.btoa("éé") is "6ek="
PASS window.btoa("\u0080\u0081") is "gIE="
PASS window.btoa("тест") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.btoa is 0
PASS typeof window.btoa is "number"
PASS typeof window.atob is "function"
PASS window.atob() threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("") is ""
PASS window.atob(null) is ""
PASS window.atob(undefined) threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob(" YQ==") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("YQ==\u000a") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("6ek=") is "éé"
PASS window.atob("6ek") is "éé"
PASS window.atob("gIE=") is ""
PASS window.atob("тест") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("z") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("zz") is "Ï"
PASS window.atob("zzz") is "Ï<"
PASS window.atob("zzz=") is "Ï<"
PASS window.atob("zzz==") is "Ï<"
PASS window.atob("zzz===") is "Ï<"
PASS window.atob("zzz====") is "Ï<"
PASS window.atob("zzzz") is "Ï<ó"
PASS window.atob("zzzzz") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("z=zz") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("=") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("==") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("===") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("====") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob("=====") threw exception Error: INVALID_CHARACTER_ERR: DOM Exception 5.
PASS window.atob is 0
PASS typeof window.atob is "number"
PASS successfullyParsed is true

TEST COMPLETE
